create useGetRandomMovies in HeroSlider component
